[None][fix] Align DSv4 cache manager helpers
