+++ tensorrt_llm/_torch/attention_backend/sparse/deepseek_v4/__init__.py
@@ -1,4 +1,4 @@
 from .cache_manager import DeepseekV4CacheManager
-from .cache_utils import DeepseekV4AttentionType
+from .deepseek_v4 import DeepseekV4AttentionType
 
 __all__ = ["DeepseekV4CacheManager", "DeepseekV4AttentionType"]
